bugfix: fixed jest listening on port indefinetly
