+++ server/tst/user.route.test.js
@@ -1,6 +1,6 @@
 const request = require('supertest');
 const mongoose = require('mongoose');
-const app = require('../server');
+const app = require('../app');
 const dotenv = require("dotenv");
 dotenv.config();
